// typings.d.ts
type Base = {
    _createdAt: string;
    _id: string;
    _rev: string;
    _type: string;
    _updatedAt: string;
    _publishedAt: string;
};

interface LayoutProps {
    className?: string;
    children: ReactNode;
}

interface Tag extends Base {
    title: string;
}

interface Category extends Base {
    title: string;
    number: number;
}
// ...
interface Author extends Base {
    image: Image;
    name: string;
    description: String;
    slug: Slug;
    socials: Socials[];
}
// ...
interface RiaProps extends Base {
    mainImage: string,
    name: string,
    decription: string,
    slug: string,
    categories: Category[];
    tags: Tag[];
    date: string,
    authors: Author[];
    body: Block[]
}

// Frequently asked questions;


interface QA {
    question: string;
    answer: string;
    category: string[];
}

interface Socials {
   
    // add more social media platforms as needed
    platform: 'twitter' | 'instagram' | 'facebook' | 'linkedin' | 'youtube' | 'twitch';
    url: string;
}
// ...
interface ImageInfo {
    image: string;
    title: string;
    location: string;
}
interface PartnerProps {
    image: string;
    title: string;
    slug: string;
}

interface InfoList  {
    id: number,
    title: string,
}